documents-master/requirements.txt
Django==4.0.1

dj-database-url==0.5.0
django-extensions==2.2.8
djangorestframework==3.13.*
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.1
drf-writable-nested==0.6.0
drf-yasg==1.20.0

whitenoise==5.0.1
sentry-sdk==1.4.0

pytest==6.2.5
pytest-django==4.5.2
factory-boy==3.2.1
pytest-cov==3.0.0
pytest-mock==3.6.1

uvicorn==0.16.0
psycopg2-binary==2.9.2
requests==2.26.*
django-environ==0.8.1

isort==5.10.1

bravado==11.0.3

cryptography==36.0.1
Pillow==9.0.1
PyMuPDF==1.19.6
